\name{Zonation5RData-package}
\alias{Zonation5RData-package}
\alias{Zonation5RData}
\title{Educational spatial data for Zonation and ZonationR}
\description{
Helper functions and static GeoTIFF data for Zonation / ZonationR teaching.
See \link{Zonation5RData-data} for sources and layer layout.
}
\keyword{package}
