% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonation5rdata_info.R
\name{zonation5rdata_info}
\alias{zonation5rdata_info}
\title{Zonation5RData Package Information}
\usage{
zonation5rdata_info()
}
\description{
Opens the package metadata (metadata.md) in a nicely rendered HTML view.
Includes layer descriptions, folder structure, and attribution information.
}
\examples{
\dontrun{
zonation5rdata_info()
}

}
